+++ crates/burn-jit/src/kernel/pool/mod.rs
@@ -1,11 +1,13 @@
 mod adaptive_avg_pool2d;
 mod avg_pool2d;
+mod avg_pool2d_backward;
 mod base;
 mod max_pool2d;
 mod max_pool2d_backward;
 
 pub(crate) use adaptive_avg_pool2d::*;
 pub use avg_pool2d::*;
+pub use avg_pool2d_backward::*;
 pub(super) use base::*;
 
 pub(crate) use max_pool2d::*;
